Updated - FlxGame: added body for handleMouseDrag and handleMouseMove. - FlxButton: correct the operator assign. Place the label on the correct place. See Fixed r6 below. - Mouse: activePointers from static to public.
Fixed
- r45
It seems that ApplicationListener::touchDown, touchUp, etc. are fired later than FlxGame::render. In the render the step() will be called which updates the inputs. That's why the position stays on the same coordinates when touch on somewhere else. To fix this I had to removed the o.screenPosition in the Mouse::update. To get the latest touch coordinates I place the screenPosition directly into Mouse::handleMouseDown, handleMouseDrag, handleMouseMove, handleMouseUp.
- r6
Having a listener for mouse events, the button label won't be updated anymore in the FlxButton:update. The get the button on the right place, the label values are assigned and also in the update-loop.

Bug
FlxButton::updateButton
When two (or more) buttons are pressed and released at the same time, one button goes into HIGHLIGHT state. That's because the button get twice into updateButton. 

1. Released two buttons simultaneous. 
2. first notifyObserver fires.
3. Everything within the observer will be fired.
4. The button checks which of the touches are in the same coordinates of the button and set it to NORMAL state.
5. second notifyObserver fires.
6. Everything within the observer will be fired.
7. The button checks which of the touches are in the same coordinates of the button and because it's already in NORMAL state it will goes into HIGHLIGHT state.

To fixed this, it could be to set the position of the touch out of the screen after justReleased I think.